refactor: Add utilities dir
