+++ tools/api-markdown-documenter/src/api-item-transforms/utilities/DocumentUtilities.ts
@@ -4,20 +4,11 @@
  */
 
 import type { ApiItem } from "@microsoft/api-extractor-model";
-import type { DocDeclarationReference } from "@microsoft/tsdoc";
-import type { Link } from "mdast";
 
-import type { ApiDocument } from "../ApiDocument.js";
-import type { Section } from "../mdast/index.js";
-import { normalizeDocumentContents } from "../mdast/index.js";
-import { resolveSymbolicReference } from "../utilities/index.js";
-
-import {
-	getDocumentPathForApiItem,
-	getLinkForApiItem,
-	shouldItemBeIncluded,
-} from "./ApiItemTransformUtilities.js";
-import type { ApiItemTransformationConfiguration } from "./configuration/index.js";
+import type { ApiDocument } from "../../ApiDocument.js";
+import { normalizeDocumentContents, type Section } from "../../mdast/index.js";
+import { getDocumentPathForApiItem } from "../ApiItemTransformUtilities.js";
+import type { ApiItemTransformationConfiguration } from "../configuration/index.js";
 
 /**
  * Creates a {@link ApiDocument} representing the provided API item.
@@ -58,37 +49,6 @@ export function createDocument(
 		contents: normalizedContents,
 		documentPath: getDocumentPathForApiItem(documentItem, config.hierarchy),
 	};
-}
-
-/**
- * Resolves a symbolic link and creates a URL to the target.
- *
- * @param contextApiItem - See {@link TsdocNodeTransformOptions.contextApiItem}.
- * @param codeDestination - The link reference target.
- * @param config - See {@link ApiItemTransformationConfiguration}.
- */
-export function resolveSymbolicLink(
-	contextApiItem: ApiItem,
-	codeDestination: DocDeclarationReference,
-	config: ApiItemTransformationConfiguration,
-): Link | undefined {
-	const { apiModel, logger } = config;
-
-	let resolvedReference: ApiItem;
-	try {
-		resolvedReference = resolveSymbolicReference(contextApiItem, codeDestination, apiModel);
-	} catch (error: unknown) {
-		logger.warning((error as Error).message);
-		return undefined;
-	}
-
-	// Return undefined if the resolved API item should be excluded based on release tags
-	if (!shouldItemBeIncluded(resolvedReference, config)) {
-		logger.verbose("Excluding link to item based on release tags");
-		return undefined;
-	}
-
-	return getLinkForApiItem(resolvedReference, config);
 }
 
 /**
