+++ tools/api-markdown-documenter/src/api-item-transforms/utilities/index.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export { createDocument, checkForDuplicateDocumentPaths } from "./DocumentUtilities.js";
+export { resolveSymbolicLink } from "./ReferenceUtilities.js";
